+++ Heimdall/Heimdall.App.WinUI/Pages/CategorySelectionPage.xaml.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.UI.Xaml.Controls;
+
+namespace Heimdall.App.WinUI.Pages;
+
+public sealed partial class CategorySelectionPage : Page
+{
+    public CategorySelectionPage()
+    {
+        InitializeComponent();
+    }
+}