+++ Heimdall/Heimdall.App.WinUI/Pages/LoadInputPage.xaml.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.UI.Xaml.Controls;
+
+namespace Heimdall.App.WinUI.Pages;
+
+public sealed partial class LoadInputPage : Page
+{
+    public LoadInputPage()
+    {
+        InitializeComponent();
+    }
+}